% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_GEDEVO}
\alias{run_GEDEVO}
\title{A simple wrapper for the GEDEVO executables}
\usage{
run_GEDEVO(args_string, verbose = F)
}
\arguments{
\item{verbose}{whether or not to print the GEDEVO output}

\item{args}{a string containing the command that would normally be
entered for running the GEDEVO executable.}
}
\description{
A simple wrapper for the GEDEVO executables
}
\examples{
\dontrun{
  cmd <- paste0(
    "./gedevo --resume test.state --create --groups DM hprd --ntw DM.ntw "
    "--ntw hprd.ntw --grsig hprd.sigs hprd --grsig DM.sigs DM --undirected "
    "--pop 1000 --maxsame 500 --no-prematch")
  run_GEDEVO(cmd)
}

}
