% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_SCOUP}
\alias{run_SCOUP}
\title{A simple wrapper for the SCOUP executable}
\usage{
run_SCOUP(method, args_string, verbose = F)
}
\arguments{
\item{method}{which executable to run, can be one of \code{"scoup"}, \code{"scoup_resume"}, \code{"sp"}, or \code{"correlation"}.}

\item{verbose}{whether or not to print the SCOUP output}

\item{args}{a string containing the command that would normally be
entered for running the SCOUP executable.}
}
\description{
A simple wrapper for the SCOUP executable
}
